Remove unnecessary arguments from non-negated cases
This is a continuation in the spirit of #104, but resolves the
awkwardness with the negative version, and brings us closer to core
Sinon messages.

To do this, we changed the nonNegatedSuffix parameter to instead take a
pair of suffixes, nonNegated and negated.  This allows us to specify %D
for the nonNegated case, which eliminates noisy duplication, while still
using %*%C for the negated case. This also matches what core Sinon
outputs in these cases.

As shorthand, we still support specifying a single value, which we
use for nonNegatedSuffix and assume an empty negatedSuffix.