Update tests for new less duplicative messages
I'm moderately happy with this. It got a little messy because now we
can't depend on the brief but duplicative listing of arguments, and have
to match against the diff view, which is colorized and has newlines and
things.

I pulled in strip-ansi@6 because it's already a transitive dependency,
and added a little wrapper function to rewrite the error messages being
thrown.

The wrapper is a little weird, but was the best solution I came up with.
Other things I tried or considered:
 - Disable color output (only possible by manipulating process.env or
   process.argv, both of which seem like a no-go)
 - Create a custom matcher that strips the message before comparing
   (throw() doesn't support matchers)
 - Matching color escapes with a RegExp (seemed way messier than this)
 - Manually try/catch, store error, assert against the error (also
   seemed messier than this)

The wrapper could be elsewhere too - it could be a wrapper that calls
expect(), for instance - but I think this is clear enough.